+++ api/models/Hosts.js
@@ -66,6 +66,11 @@ module.exports = {
             type: 'boolean',
             defaultsTo: false
         },
+        
+        accountability: {
+            type: 'boolean',
+            defaultsTo: false
+        },
 
         webmessages: {
             type: 'boolean',